+++ memiavl/multitree.go
@@ -358,6 +358,10 @@ func (t *MultiTree) CatchupWAL(wal *wal.Log, endVersion int64) error {
 }
 
 func (t *MultiTree) WriteSnapshot(dir string, wp *pond.WorkerPool) error {
+	return t.WriteSnapshotWithContext(context.Background(), dir, wp)
+}
+
+func (t *MultiTree) WriteSnapshotWithContext(ctx context.Context, dir string, wp *pond.WorkerPool) error {
 	if err := os.MkdirAll(dir, os.ModePerm); err != nil {
 		return err
 	}
@@ -368,7 +372,7 @@ func (t *MultiTree) WriteSnapshot(dir string, wp *pond.WorkerPool) error {
 	for _, entry := range t.trees {
 		tree, name := entry.Tree, entry.Name
 		group.Submit(func() error {
-			return tree.WriteSnapshot(filepath.Join(dir, name))
+			return tree.WriteSnapshotWithContext(ctx, filepath.Join(dir, name))
 		})
 	}